+++ lib/node.cc
@@ -1,7 +1,8 @@
 //
 // This file is part of khmer, http://github.com/ged-lab/khmer/, and is
 // Copyright (C) Michigan State University, 2009-2013. It is licensed under
-// the three-clause BSD license; see doc/LICENSE.txt. Contact: ctb@msu.edu
+// the three-clause BSD license; see doc/LICENSE.txt.
+// Contact: khmer-project@idyll.org
 //
 
 #include "node.hh"
@@ -106,7 +107,7 @@ Kmer Node::makeNextKmer(unsigned char forward, char b)
 
 Node::Node(Node* _parent,
            char _emission,
-           unsigned int _stateNo,
+           unsigned long _stateNo,
            char _state,
            Kmer _kmer) : kmer( _kmer)
 {
@@ -136,8 +137,8 @@ std::queue<Node*> Node::enumerate(CountingHash* ch,
 {
     std::queue<Node*> ret;
 
-    int index;
-    int remaining;
+    size_t index;
+    size_t remaining;
     double bestMatch = sm->score('A','A');
     double errorOffset = 20.0;
